CI test more and cache image (#92)

+++ scripts/sdk_installers/install_zisk_sdk.sh
@@ -59,21 +59,23 @@ else
     exit 1
 fi
 
-# Step 3: Build cargo-zisk-gpu from source with GPU features enabled
-TEMP_DIR=$(mktemp -d)
-git clone https://github.com/0xPolygonHermez/zisk.git --single-branch --branch "v$ZISK_VERSION" "$TEMP_DIR/zisk"
-cd "$TEMP_DIR/zisk"
-cargo build --release --features gpu
-cp ./target/release/cargo-zisk "${HOME}/.zisk/bin/cargo-zisk-gpu"
-cp ./target/release/libzisk_witness.so "${HOME}/.zisk/bin/libzisk_witness_gpu.so"
-rm -rf "$TEMP_DIR"
+# Step 3: Build cargo-zisk-gpu from source with GPU features enabled (skip if in CI)
+if [ -z $CI ]; then
+    TEMP_DIR=$(mktemp -d)
+    git clone https://github.com/0xPolygonHermez/zisk.git --single-branch --branch "v$ZISK_VERSION" "$TEMP_DIR/zisk"
+    cd "$TEMP_DIR/zisk"
+    cargo build --release --features gpu
+    cp ./target/release/cargo-zisk "${HOME}/.zisk/bin/cargo-zisk-gpu"
+    cp ./target/release/libzisk_witness.so "${HOME}/.zisk/bin/libzisk_witness_gpu.so"
+    rm -rf "$TEMP_DIR"
 
-echo "Checking for cargo-zisk-gpu CLI tool..."
-if cargo-zisk-gpu --version; then
-    echo "cargo-zisk-gpu CLI tool verified successfully."
-else
-    echo "Error: 'cargo-zisk-gpu --version' failed." >&2
-    exit 1
+    echo "Checking for cargo-zisk-gpu CLI tool..."
+    if cargo-zisk-gpu --version; then
+        echo "cargo-zisk-gpu CLI tool verified successfully."
+    else
+        echo "Error: 'cargo-zisk-gpu --version' failed." >&2
+        exit 1
+    fi
 fi
 
 # Step 4: Make sure `lib-c`'s build script is ran.
